src/requirements.txt
pandas==2.3.3
numpy==2.3.5
pip==24.3.1
pillow==12.0.0
python-dotenv==1.2.1
psycopg==3.3.2
psycopg2
sqlalchemy
matplotlib==3.10.8
joblib==1.5.3

requests==2.32.5
celery==5.6.0

seaborn==0.13.2
imbalanced-learn==0.14.0
xgboost==3.1.2
scipy==1.16.3
tornado==6.5.4
scikit-learn==1.5.2